requirements.txt
# Git integration
gitpython==3.1.40

# Vector database
chromadb==0.4.22

# LLM and embeddings
google-generativeai==0.3.2
sentence-transformers==2.2.2

# Data validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Web UI
gradio==4.12.0

# Code syntax highlighting
pygments==2.17.2

# Markdown processing
markdown2==2.4.12

# Code parsing
tree-sitter==0.20.4
tree-sitter-languages==1.10.2

# Environment variables
python-dotenv==1.0.0

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.23.0